added forgotten changes

+++ T2408_ATtiny412_edog/eep.cpp
@@ -8,7 +8,7 @@
 
 typedef struct
 {
-  uint8_t  addr;
+  int  addr;
   uint8_t  save_cntr ;
 } eep_st;
 
@@ -48,19 +48,19 @@ void eep_req_save(uint8_t eep_indx)
   if(eep[eep_indx].save_cntr == 0) eep[eep_indx].save_cntr = EEP_CNTR_START;
 }
 
-void eep_save_u8(eeprom_index_et indx, uint8_t offs, uint8_t u8 )
+void eep_save_u8(eeprom_index_et indx, int offs, uint8_t u8 )
 {
   EEPROM.write(eep[indx].addr + offs, u8);
 }
 
-uint8_t eep_load_u8(eeprom_index_et indx, uint8_t offs  )
+uint8_t eep_load_u8(eeprom_index_et indx, int offs  )
 {
   return EEPROM.read(eep[indx].addr + offs);
 }
 
 void eep_save_array(eeprom_index_et indx, uint8_t bytes, uint8_t *u8_arr )
 {
-  uint8_t addr = eep[indx].addr;
+  int addr = eep[indx].addr;
   for (uint8_t i = 0; i < bytes; i++)
   {
     EEPROM.write(addr++, u8_arr[i]);
@@ -69,8 +69,8 @@ void eep_save_array(eeprom_index_et indx, uint8_t bytes, uint8_t *u8_arr )
 
 void eep_load_array(eeprom_index_et indx, uint8_t bytes, uint8_t *u8_arr )
 {
-  uint8_t addr = eep[indx].addr;
-  for (uint8_t i = 0; i < bytes; i++)
+  int addr = eep[indx].addr;
+  for (int i = 0; i < bytes; i++)
   {
     u8_arr[i] = EEPROM.read(addr++);
   }
